DEV: add skip feature if dotfile already in destination directory

+++ syndot/__main__.py
@@ -44,19 +44,21 @@ elif args.command == 'link':
     for file in target_files:
         source_file_path = os.path.join(os.path.expanduser(source), file)
         destination_file_path = os.path.join(os.path.expanduser(destination), file)
-        if args.backup:
-            shutil.copy(source_file_path, destination_file_path)
-            os.rename(source_file_path, os.path.join(os.path.expanduser(source), file + '.bak'))
-        else:
-            shutil.move(source_file_path, destination_file_path)
-        os.symlink(destination_file_path, source_file_path)
+        if not os.path.exists(destination_file_path):
+            if args.backup:
+                shutil.copy(source_file_path, destination_file_path)
+                os.rename(source_file_path, os.path.join(os.path.expanduser(source), file + '.bak'))
+            else:
+                shutil.move(source_file_path, destination_file_path)
+            os.symlink(destination_file_path, source_file_path)
 
     for directory in target_directories:
         source_directory_path = os.path.join(os.path.expanduser(source), directory)
         destination_directory_path = os.path.join(os.path.expanduser(destination), directory)
-        if args.backup:
-            shutil.copytree(source_directory_path, destination_directory_path)
-            os.rename(source_directory_path, os.path.join(os.path.expanduser(source), directory + '_bak'))
-        else:
-            shutil.move(source_directory_path, destination_directory_path)
-        os.symlink(destination_directory_path, source_directory_path)
+        if not os.path.exists(destination_directory_path):
+            if args.backup:
+                shutil.copytree(source_directory_path, destination_directory_path)
+                os.rename(source_directory_path, os.path.join(os.path.expanduser(source), directory + '_bak'))
+            else:
+                shutil.move(source_directory_path, destination_directory_path)
+            os.symlink(destination_directory_path, source_directory_path)
